Add and modify some plot
I added a countplot with the best clients for know amount of transations for each one.
I modified the countplot where show the distribution of ages for a countplot with three bar for easy visualization.
I added two boxplots with gender by sales quantity and gender by total amount.
And finally I added a headmap for know the correlation on the dataset.